Comment out line that was erroring; should get VIVO eyes on it
For some reason, two hashes do not match. I think it may be because of
the type annotation on the string, which is always included with
MarkLogic but perhaps not with Jena, and the way VIVO handles it.

I do not know why this code is needed. The back button seems to work
fine for me. Famous last words. :-)